Format / refactor code to PSR standards, introduce exceptions, update documentation
Code was reformatted to match PSR standards. Exceptions were introduced to assist with error handling. API responses were modeled in a response object to assist with property access and reduce potential for typos and errors. A composer.json file was introduced to allow installation via Composer (and hopefully later Packagist). Documentation was updated to match.